requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncer==0.0.7
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext[openai]==0.7.5
bidict==0.23.1
boto3==1.42.25
botocore==1.42.25
certifi==2026.1.4
chainlit==2.3.0
charset-normalizer==3.4.4
chevron==0.14.0
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.115.14
filetype==1.2.0
FLAML==2.3.6
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.12.0
jmespath==1.0.1
Lazify==0.4.0
literalai==0.1.103
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
plotly==6.5.1
propcache==0.4.1
protobuf==5.29.5
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.18
python-socketio==5.16.0
pytz==2025.2
pywin32==311
regex==2025.11.3
requests==2.32.5
s3transfer==0.16.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.41.3
syncer==2.0.3
termcolor==3.1.0
tiktoken==0.12.0
tomli==2.4.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uptrace==1.39.0
urllib3==1.26.20
uvicorn==0.39.0
watchfiles==0.20.0
wrapt==1.17.3
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
langchain==0.3.7
langchain-core==0.3.18
langchain-openai==0.3.3